Combat pas terminé

+++ C++ard game/Draw.hpp
@@ -50,7 +50,7 @@ public:
 	};
 	std::vector<Card*> complete_shop_possible_card(std::vector<Card*> possible_card, std::vector<Card*> pShop) {
 		std::vector<Card*> temp = pShop;
-		int howmany = 3 - pShop.size();
+		int howmany = 3 - (int)pShop.size();
 		for (int i = 0; i < howmany; i++) {
 			int random = rand() % possible_card.size();
 			Card* card = new Card(*possible_card[random]);
